+++ src/cppNGS/BamReader.h
@@ -10,6 +10,7 @@
 
 #include "QHash"
 
+#include "RefGenomeService.h"
 #include "htslib/sam.h"
 #include "htslib/cram.h"
 
@@ -282,7 +283,7 @@ class CPPNGSSHARED_EXPORT BamReader
 		BamReader(const QString& bam_file);
 		//CRAM Constructor with explicit reference genome
 		//reference genome is mandatory for CRAM support
-		BamReader(const QString& bam_file, const QString& ref_genome);
+        BamReader(const QString& bam_file, QString ref_genome);
 
 		//Destructor
 		~BamReader();
@@ -338,7 +339,7 @@ class CPPNGSSHARED_EXPORT BamReader
 		//Releases resources held by the iterator (index is not cleared)
 		void clearIterator();
 		void verify_chromosome_length(const QString& ref_genome);
-		void init(const QString& bam_file, const QString& ref_genome = QString());
+        void init(const QString& bam_file, QString ref_genome = QString());
 
 		//"declared away" methods
 		BamReader(const BamReader&) = delete;